pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loom"
version = "0.1.0"
description = "Runtime scaffolding for faithful long-horizon persona simulation via API-accessed models"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
loom = "loom.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live: marks tests that make real API calls (deselect with '-m \"not live\"')",
]
